use getNumRows instead of getLastRow
