+++ src/server/accounts/account-utils.test.ts
@@ -1,10 +1,11 @@
 import { AccountUtils, isNumeric } from './account-utils'
-import { getSheetById } from '../globals'
+import { FireSpreadsheet, getSheetById } from '../globals'
 import { RangeMock, SheetMock } from '../../../test-setup'
 import { SOURCE_SHEET_ID } from '@/common/constants'
 
 describe('Utility tests', () => {
   beforeEach(() => {
+    vi.clearAllMocks()
     AccountUtils.resetStaticCache()
   })
 
@@ -47,6 +48,40 @@ describe('Utility tests', () => {
     expect(() => AccountUtils.getBalance('openbank')).toThrow(
       'Account \'openbank\' not found',
     )
+  })
+
+  test('getBalance should call getRangeByName every time (Baseline)', () => {
+    const getRangeByNameSpy = vi.spyOn(FireSpreadsheet, 'getRangeByName')
+
+    // Mock the spreadsheet response
+    RangeMock.getValues.mockReturnValue([
+      ['N26', 'DB123456789', '302.80'],
+      ['Openbank', 'BANK123456789', '400'],
+    ])
+
+    // Call getBalance 3 times
+    AccountUtils.getBalance('n26')
+    AccountUtils.getBalance('n26')
+    AccountUtils.getBalance('n26')
+
+    // After optimization, it should be called only once
+    expect(getRangeByNameSpy).toHaveBeenCalledTimes(1)
+  })
+
+  test('getBankAccounts and getBalance should both trigger range lookups independently', () => {
+    const getRangeByNameSpy = vi.spyOn(FireSpreadsheet, 'getRangeByName')
+
+    RangeMock.getValues.mockReturnValue([
+      ['N26', 'DB123456789', '302.80'],
+      ['Openbank', 'BANK123456789', '400'],
+    ])
+
+    AccountUtils.getBankAccounts()
+    AccountUtils.getBalance('n26')
+
+    // After optimization, they share the same cache/fetch method.
+    // So it should be called only once.
+    expect(getRangeByNameSpy).toHaveBeenCalledTimes(1)
   })
 })
